Corrected interface name Moved large function inside component to utils file
