+++ openmetadata-ui/src/main/resources/ui/src/components/DataAssets/DataAssetsHeader/DataAssetsHeader.interface.ts
@@ -61,7 +61,7 @@ export type DataAssetsWithoutServiceField =
   | MetadataService
   | StorageService;
 
-export type DataAssetsWithoutFollowersField = Exclude<
+export type DataAssetsWithFollowersField = Exclude<
   DataAssetsType,
   DataAssetsWithoutServiceField | Database | DatabaseSchema
 >;
